+++ backend/controllers/task.js
@@ -2,8 +2,17 @@ const Task = require("../models/tasks");
 
 // Create a new task
 exports.createTask = (req, res, next) => {
-    const task = new Task(req.body);
-    
+    let task;
+    try {
+        task = new Task(req.body);
+    } catch (err) {
+        res.status(400).json({
+            message: "Parameters did not match Model",
+        });
+    }
+    if(req.TokenUserId && !task.taskAdministrators.includes(req.TokenUserId)) {
+        task.taskAdministrators.push(req.TokenUserId);
+    }
     task.save()
     .then((result) => {
         res.status(201).json({
@@ -22,19 +31,21 @@ exports.createTask = (req, res, next) => {
 // Update a task by ID
 exports.updateTask = (req, res, next) => {
     const taskId = req.params.id;
-
+    
     Task.findById(taskId)
     .then((task) => {
         if (!task) {
             throw new Error("Task not found");
         }
 
-        // Add fields that are allowed to be updated
-        task.taskName = req.body.taskName;
-        task.taskAdministrators = req.body.taskAdministrators;
-        task.employees = req.body.employees;
-        task.status = req.body.status;
-        // ...other fields you might want to update
+        if(req.TokenUserId && !task.taskAdministrators.includes(req.TokenUserId)) {
+            throw new Error("Invalid Credentials");
+        }
+
+        if (req.body.taskName !== undefined) task.taskName = req.body.taskName;
+        if (req.body.taskAdministrators !== undefined) task.taskAdministrators = req.body.taskAdministrators;
+        if (req.body.employees !== undefined) task.employees = req.body.employees;
+        if (req.body.status !== undefined) task.status = req.body.status;
 
         return task.save();
     })
@@ -94,21 +105,29 @@ exports.getTaskById = (req, res, next) => {
 exports.deleteTask = (req, res, next) => {
     const taskId = req.params.id;
 
-    Task.findByIdAndRemove(taskId)
+    Task.findById(taskId)
     .then((task) => {
         if (!task) {
             throw new Error("Task not found");
         }
 
+        // Check if the user has the necessary credentials to delete the task
+        if (req.TokenUserId && !task.taskAdministrators.includes(req.TokenUserId)) {
+            throw new Error("Invalid Credentials");
+        }
+
+        // If the user has the credentials, proceed with task deletion
+        return Task.findByIdAndRemove(taskId);
+    })
+    .then(() => {
         res.json({
             message: "Task deleted successfully",
-            task: task.toObject(),
         });
     })
     .catch((err) => {
         console.log(err);
         res.status(404).json({
-            message: err.message || "Task not found!",
+            message: err.message || "Task not found or invalid credentials!",
         });
     });
 };
